This library contains a small library for pseudodynamic arrays.
It allows to create such arrays, then fill them with values.
The values can be mixed. This means the arrays can hold integer and string values at the same time.

There are some small and rudimentary benchmark information for OS windows version.
The mac and linux version have no benchmark and are not tested on their specific OS.