Further performance improvements are made in IWDEE_SEQUENCE_TEXTPOS_AND_TITLE.
The input array check is commented out, which provides almost the same performance when the sanity check is disabled.

what remains is sanity check for the identity of title and strref in the LUA files.

The risk of duplication of the input array, i.e. that within the input array there might be one or more identical strrefs or titles, should be checked before installation, like with Excel or Libre Office Calc.

As presumably these offer better time performance.

This means that the parameter to disable sanity check is retained, for whatever reason someone wants to use the commented out code in IWDEE_SEQUENCE_TEXTPOS_AND_TITLE.